v0.1.4
Updating and future-proving the messages structure for planned operations.
ClientConnectorImpl now supports connector-wise configuration of "required confirmations" and specifying the port for connector to listen.
Better (more friendly) interface to Keygen API in the connector.
Backend: updating the logic to track the currencies (tokens) even if they are not supported yet.
Backend: server-side improvements for launching multiple components at once (useful for development and debugging).