v0.1.5
Observer now provides the API to get the list of transfers:
from/to specific block, with at least specified number of confirmations,
from/to specific (tracked) addresses, and with other features.
The result list of transfers can be extra-filtered and is indexed by any senders and receivers.
Also: CherryGardener Connector CLI now supports the conf-file to store the connection details.