v0.9.1
Store the transfer operation gas limit for all supported currencies. For verified currencies, that's a mandatory field.
More optimizations for message serialization (in particular, to prepare the structures for sending the transactions, and to transfer the transferGasLimit).
Nicer API (EAFP-way) to get the response payload, if all you want is the successful payload - so you get it, and handle the exception otherwise.